+++ src/Encoder.cpp
@@ -1,7 +1,7 @@
 #include "Encoder.h"
 #include <Arduino.h>
 
-void Dn_Encoder::init(int s1Pin, int s2Pin, int keyPin)
+void Dn_Encoder::init(uint8_t s1Pin, uint8_t s2Pin, uint8_t keyPin)
 {
     this->s1Pin = s1Pin;
     this->s2Pin = s2Pin;
